chore: abstract away secret watcher
